Implemented methods of TransactionFullLoader and Abstract parent for all the loaders.
